pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "symphony"
version = "0.1.0"
description = "Symphony - Agent Orchestration System (Python Edition)"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
authors = [
    {name = "Symphony Contributors"},
]
keywords = ["agent", "orchestration", "linear", "ai", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core
    "agentscope>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    
    # HTTP & API
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    
    # Template & Parsing
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    
    # Scheduling
    "apscheduler>=3.10.0",
    
    # Terminal UI
    "rich>=13.9.0",
    
    # Logging
    "structlog>=24.4.0",
    
    # File watching
    "watchfiles>=0.24.0",
    
    # SSH (optional but recommended)
    "asyncssh>=2.17.0",
    
    # CLI
    "click>=8.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "orjson>=3.10.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.0",
    "pytest-timeout>=2.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "types-pyyaml>=6.0.0",
    "types-python-dotenv>=1.0.0",
]

[project.scripts]
symphony = "symphony.cli:main"

[project.urls]
Homepage = "https://github.com/openai/symphony"
Documentation = "https://github.com/openai/symphony/tree/main/symphony.py"
Repository = "https://github.com/openai/symphony"
Issues = "https://github.com/openai/symphony/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/symphony"]

[tool.hatch.metadata]
allow-direct-references = true

# Ruff Configuration
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q",   # flake8-quotes
    "FURB",# refurb
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["symphony"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy Configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = ["agentscope.*", "asyncssh.*"]
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "-ra",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
    "linear: Tests requiring Linear API",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src/symphony"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false
fail_under = 80

[tool.coverage.html]
directory = "htmlcov"
